Address PR comments

+++ backend/canisters/group_index/api/src/updates/set_max_concurrent_group_canister_upgrades.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub value: usize,
+    pub value: u32,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
